[ft_printf] fix print hex!!

+++ 01_ft_printf/src/util/ft_putchar_n_out.c
@@ -6,13 +6,13 @@
 /*   By: jaeskim <jaeskim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/19 17:18:36 by jaeskim           #+#    #+#             */
-/*   Updated: 2020/10/21 14:21:03 by jaeskim          ###   ########.fr       */
+/*   Updated: 2020/10/26 22:26:44 by jaeskim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf_util.h"
 
-void	ft_putchar_n_out(char **out, size_t n, int c)
+void	ft_putchar_n_out(char **out, int n, int c)
 {
 	while (n-- > 0)
 		ft_putchar_out(out, c);
